Update patch set 2
Patch Set 2:

In the bug, it was proposed that since dalvik.system.VMStack.getCallingClassLoader is @FastNative, it is broken under certain configurations (such as with generic JNI).

So I wrote this test (and ran it in that mode) to check whether or not that's the case. It seems to work fine.

Would you like me to change this somehow? Is this missing anything?

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0