Update patch set 2
Patch Set 2:

I added some magic to the run script to force core.oat not to be used and it was still passing.

And I filed a bug b/36136075 here because it's probably a good idea to test everything in ART without a core.oat and not just do it for a single test. Especially when the arguments to add are non-obvious.

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0