Improve splitting/spilling position choice.
Improve splitting/spilling position choice for a particular case
when "active" interval is split exactly at the start position of
"current" interval regardless its previous uses, loop structure,
etc. The patch fixes the problem when a value is reloaded just
before the back edge of a loop and has no uses inside and thus is
redundant. This happens when interval is split in the middle of the
loop however split position can be hoisted out of it. The case is
exploited only when it is safe to change the split point and it
doesn't break any invariants of linear scan algorithm.

On ARM (32bit) the patch:
 * reduces the number of spill/fill loads by up to 23%.
 * brings up to 15% (LoopAtom) performance improvement.
 * brings up to 2.5% .text size reduction.

Test: mma test-art-host && mma test-art-target

Change-Id: I6a0c776de41785a7734f3cf0b68c46a3a76be666