Tweak ProfileSaver saving strategy
To minimize the I/O, we used to store the number of methods/classes we
last saves globally in the profiler saver. This is no longer viable once
we want to track secondary dex files for profiling because each file
might saved a different set of methods.

To make sure we do not miss data for secondary dex profiles,
store the number of last saved methods separetely for each profile file.

Test: test-art-host
Bug: 26719109
Change-Id: I3a657f63d26f68c7ca83a754f6e4aa2c9d946176