Update patch set 6
Patch Set 6: Code-Review+1

This clearly fixes a memory model issue. I don't have a strong opinion on the other issues under discussion.

IIUC, it seems this was and remains much slower than necessary for copying byte arrays, since we copy a byte at a time? We should presumably be using an approximate equivalent of memcpy to actually do the work for large arrays. Or does that case not get here?

Discussing this with Hiroshi, it also wasn't clear to us that the read barrier code is needed in the primitive array case. Again, that's a potential preexisting issue, not a critique of this CL.

Patch-set: 6
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1