Update patch set 1
Patch Set 1:

This fixes the slow path omission of not saving full state ("default" ABI only saves 64-bit for all xmm regs). An example of how it looks is shown below. Note that this is only done for xmm regs that are live on the path, which is usually not too many. Also note that spilling in parallel move resolver is still TBD.


      0x00001174:           660F114C2414               movupd [esp + 20], xmm1
      0x0000117a:           660F11542424               movupd [esp + 36], xmm2
      0x00001180:         64FF15B0020000               call fs:[0x2b0]  ; pTestSuspend
        StackMap [native_pc=0x1187] [entry_size=0x11 bits] (dex_pc=0x1, native_pc_offset=0x147, dex_register_map_offset=0xffffffff, inline_info_offset=0xffffffff, register_mask=0x1, stack_mask=0b)
      0x00001187:           660F104C2414               movupd xmm1, [esp + 20]
      0x0000118d:           660F10542424               movupd xmm2, [esp + 36]
      0x00001193:             E944FFFFFF               jmp -188 (0x000010dc)

Patch-set: 1
Reviewer: Gerrit User 1074526 <1074526@85c56323-6fa9-3386-8a01-6480fb634889>