Update patch set 2
Patch Set 2:

Hmm. This is my WIP getting the save/restore to work in quick entry points. However, it crashes pretty much everything. I am guessing it is not so simple after all.

Sending this out to get input (perhaps there is only a few places where this can be easily fixed). But if not, perhaps the other way is the way to go.

Patch-set: 2
Reviewer: Gerrit User 1074526 <1074526@85c56323-6fa9-3386-8a01-6480fb634889>