Update patch set 3
Patch Set 3:

Ok, after some more digging I am starting to think this approach is way too risky. The long jumps are implemented around the same mechanism, and I am not sure it is a good idea to break that open.

Nicolas, shall we go back to plan B (or plan A depending on your view), and just save the live xmm registers along the suspend check? That was more or less a 4 line solution that was low risk and high gain :-)

Patch-set: 3
Reviewer: Gerrit User 1074526 <1074526@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0