Remove test logic from Android.run-test.mk
As python testrunner replaces it.

This change removes the following targets:
test-art-host-run-test-[variant]

The following targets are still supported:
test-art-host-run-test
test-art-target-run-test
test-art-run-test

Each of the above targets invoke the new testrunner.
They don't pass any concurrency value to the testrunner
and the testrunner itself decides the concurrency
depending upon the type of tests.

Test: make -j test-art-host-run-test
      make -j test-art-target-run-test
      make -j test-art-run-test
      ./art/test/testrunner/testrunner.py --host
      ./art/test/testrunner/testrunner.py --target

Change-Id: I64a87d0f2e62c29b4d9c14527f9a83810ae7edc7