Add call stack tracker and use it for dex file accesses
Track the and count unique stack traces for accessors to DexFile
data. This can be useful to reduce the amount of dex file accesses
that pull in code pages.

The feature is enabled by the following option:
-Xtrackdexaccesses:<dexes>

The dex file is tracked if there is a substring in the tracking
string. Multiple dexes are supported if separated by commas.
Example:
/data/app/com.google.android.apps.maps,/data/app/com.google.android.youtube

The data is dumped in SIGQUIT if there is any collected.

Currently tracks DexFile calls to GetCodeItem, GetTypeId, GetFieldId,
GetMethodId, and GetStringId.

Bug: 35800981
Bug: 36457259

Test: test-art-host
Test: adb shell setprop dalvik.vm.extra-opts '"-Xtrackdexaccesses:/data/app/com.google.android.apps.maps"'
Test: adb shell kill -s QUIT `pid com.google.android.apps.maps`

Change-Id: I7a134fdeb33478a15df00b7cc6c7ed922ee1ae68