Recompute Dex Checksum before passing it to agents
Some internal tools might modify dex files. These tools generally do
not recompute or fix up the dex file checksum for a variety of
reasons. We recompute the checksum for the dex file before passing it
to agents to hide this implementation detail.

TODO: Figure out which dex file modifications are being made to
debuggable apps.

Bug: 31455788
Test: ./test.py --host -j40
Change-Id: Ibd75e21d6d9f19d9bb125a43a0f54262724e894c