Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> The flag should definitely be atomic.
> 
> Thinking a bit more about this, I unfortunately think that we do need a seq_cst load in the fast path. Otherwise consider the Dekker's example variant we discussed earlier:
> 
> Thread 1: thread2.interrupt(); interrupted();
> 
> Thread 2: thread1.interrupt(); interrupted();
> 
> If interrupted() uses an initial memory_order_relaxed load, it's possible that neither thread will notice it has been interrupted. As Igor explained, this contradicts the spirit of the JLS. Whether it contradicts the letter of the JLS is unclear.
> 
> Probably this still helps noticeably?

Can we still do this with a relaxed load on the fast path (when there's no interrupt)?

What if we put 2x seq_cst fence into #isInterrupted and #interrupt (at the beginning and end of function), then I think it solves your example above because the fence should prevent reordering of the relaxed load. Perhaps it doesn't even need the 2x fences in the other calls (but with 2, it seems like it would work and even prevent IRIW).

I am guessing the common case is just to call #interrupted over and over again, presumably to check if the exception should be thrown. This way the less common case is somewhat more expensive and the fast case is very fast.

Patch-set: 3
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>