Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I'm not sure this should duplicate logic from the makefile, it will make maintaining later more difficult.
> 
> Can you refactor the makefile to call this new script?

Not easily since this script specifically only calls it for changed files.

Honestly I don't think it will be much of an issue. The filter flags haven't changed since '13 and the --quiet is non-essential.

The only thing that could change much is the ignored files but those are only for external files that we need which are added extremely rarely. I don't think refactoring the makefile and the runner-script is worth the effort.

Patch-set: 3
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>