Ensure invoking obsolete methods throws errors.
We never actually checked that a method being invoked was not
obsolete. This adds a check in ArtMethod::Invoke code. Getting into
this state is extreemly difficult since one can only get an jmethodID
to an obsolete method by snatching it off the stack (or by inspecting
internal runtime data). From there normally invoking it will do lookup
on the receiver which will get you the original version.

Bug: 36867251
Bug: 31455788
Test: ./test.py --host -j40

Change-Id: I2ca0503966a4e3de18dd89cb7ff224eba1459b49