Use cpu-info for feature set when running JIT.
Rationale:
Under JIT, the current CPU is the one we are compiling
for, so we can safely extract any supported features
(if nothing is given explicitly as options). This enables
more target specific optimizations, such as SIMDization.

Sample golem improvements (JIT x86_64):
(due to SIMDization, reversi due to popcount exploit)
  Fannkuch           7.956%
  Sieve             21.241%
  Linpack           42.962%
  LU                52.943%
  ReversiEvalBench 112.019%

Bug: 30933338
Test: test-art-host

Change-Id: If16160b0c699d9bc95c8e79caf8d1f163498b848