optimizing: constructor fence redundancy elimination - remove dmb after LSE
Part one of a few upcoming CLs to optimize constructor fences.

This improves load-store-elimination; all singleton objects will
have their associated constructor fence removed (e.g. if the allocation
is removed, so is the fence).

This change is enabled by tracking the "this" object associated with the
constructor fence. Once all real uses (other than the fence itself)
disappear, it is safe to also remove the fence.

Bug: 36656456
Test: art/test.py --host
Change-Id: I05659ab07e20d6e2ecd4be051b722726776f4ab1