Better encapsulate dalvik cache file naming scheme.
The goal of this CL is to make it easier to change the dalvik cache file
naming scheme in the future, in case such a change is deemed necessary.

This is acheived by:
1. Reusing GetDalvikCacheFilename in patchoat rather than duplicating
the logic for the naming scheme there

2. Changing the encoding of the boot class path for multi-image so
that it does not rely on '@' (from the current dalvik cache file
naming scheme) acting as a directory separator.

In the case of multi-image, a separate image and oat file is generated
for each element on the boot class path. These files are named using a
combination of the primary image name (which is a special case) and
the boot class path element they correspond to. For example, given a
boot class path of 'foo.jar', 'bar.jar', and 'sludge.jar' and a primary
image name of 'boot.art', the rest of the image names are
'boot-bar.art' and 'boot-sludge.art'.

Previously the boot class path was stored as a string in the primary
image's oat file using a colon-separated list of file names consisting
of the path portion of the boot class path element and some part of
the name portion of the image name depending on '@' characters in the
name. This CL changes how the boot class path is stored in the primary
image's oat file to be a colon-separated list of the boot class path
elements with the path portion stripped. For example, the boot path
class would now be stored as "foo.jar:bar.jar:sludge.jar", which is
sufficient to reconstruct the image names given the name of the
primary image.

Bug: 37207894
Test: m test-art-host
Test: run_build_test_target.py art-no-image

Change-Id: I86dff6c4b0dc9a5f2f3d87068ebf7761148ca446