Update patch set 1
Patch Set 1:

I find this naming scheme, while simpler, a lot uglier. Having the test configuration be a name suffix for the first, but infix for the others looks ****.

Please detail "simpler encoding" in the commit message (it also has not a lot to do with "better encapsulate," but that's a different topic).

Patch-set: 1
CC: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>