Update patch set 2
Patch Set 2:

? How is it being used APCT? Is there source? Design doc?


If they are running for APCT I think the right future solution is to move to using run_build_test_target.py and then have the logic there so we/they can refactor it later and not break things.

That's what we have done for go/ab, are in the process of doing for go/lem, and we should do it for everything else [in the testing infrastructures] as well so we can limit the amount of things breaking every time we change our command line interfaces.

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>