Update patch set 4
Patch Set 4:

(8 comments)

> Could you split this into 2 CLs?:
> 1) introduce HConstructorFence. This amounts to refactoring,
> 2) actually remove the fences in LSE.

Discussed offline. 

This is exactly what I wanted to do to begin with. I had only (1) done.

With only (1), this is what happens:

a) it doesn't remove new-instance when it should because constructor fence keeps it alive [checker tests fail]
b) you make it ignore the constructor fence and then other things DCHECK-fail or crash because the inputs to the constructor fence are gone [crashes when dex2oat runs]
c) you end up writing a bunch of code to remove inputs in LSE explicitly... and we're basically at where it is today

Patch-set: 4
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0