ART: Refactor run-test 913
Remove references to ART. Filter roots from the JIT. Canonicalize
some thread IDs. Move the test to its own thread, and filter
stack-locals of other threads.

These changes ensure that the test is less dependent on the main
thread and its environment, which is required to use it in CTS.

Also fix a reporting issue for roots.

Bug: 32072923
Test: art/test/testrunner/testrunner.py --host -t 910
Change-Id: I8480ba7751fb6420c256db87cba11b8a65e25ea5