Update patch set 1
Patch Set 1:

How does this happen? I thought dex2oat supports all the ISA variants that we have dalvikvm on-device for?

(and in particular if we have dalvikvm 64 -> there is also dalvikvm 32) is there ever a case where we only have dalvikvm 64 and not 32?

Patch-set: 1
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>