Update patch set 6
Patch Set 6:

> I'm really thinking about a -1.

Well, that seems a bit overly harsh. I agree there is a certain inelegance in recognizing idioms, but the buildings blocks here are not *that* bad, and can be easily adapted to the type system changes we may do in the future (leaving the rest of the vectorization code the same).

Patch-set: 6
Reviewer: Gerrit User 1074526 <1074526@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0