Update patch set 1
Patch Set 1:

I think this is a reasonable fix. But I don't see a reason not to just replace the loop with, say snooze(2000). AFAICT, the only reason for the loop was to catch the case in which we time out the finalizer immediately, rather than waiting for 10 seconds. Just sleeping seems to be a much more robust (though presumably still very slightly flakey) way to do that.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>