Update patch set 1
Patch Set 1:

> Could you also work to setup a run-test to check this scenario?
 > 
 > something local like , run-test --strip could do.
 > 
 > Can be done in a follow up.

Actually this turns out to be harder than I expected. I wanted to include it in this change, but in order to properly test this, we need to generate an out of date oat file with an up-to-date vdex file. Probably a bit too much for run-tests, and that probably falls in the camp of PackageManager dexopt tests.

For now, let's live with oat_file_assistant_test.

Patch-set: 1
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>