Resolve tracked location in profile saver if needed.
The profile saver takes the dex location it has to track as they are. On
the built bot the locations are symlinked, but the dex file are compiled
with absolute path so there's a missmatch and things may not get
profiled making test 595-profile-saving to fail.

Note that this is a valid use case in production as well. It may happen
for secondary dex files which can be loaded via symlinks.

Test: test/run-test --output-path /tmp/link --always-clean --host --prebuild --interpreter --no-relocate --runtime-option -Xcheck:jni --no-image --64 595-profile-saving
       test/run-test  --output-path /tmp/link --always-clean --host
--no-prebuild --jit --no-relocate --runtime-option -Xcheck:jni
--runtime-option -verbose:profiler --64  595-profile-saving

Bug: 36605597
Change-Id: I0126d5255fd663d232a3f7fc50ce7d715072872a