Update patch set 1
Patch Set 1:

> I'd suggest the other way :) Isn't the entrypoint change just a one line change (ie removing the fence) per arch?

Depends on whether or not you care about getting code size back.

The _initalized_ version of the art_quick_alloc wouldn't have the DMB, but if we duplicate it and call it _initialized_with_dmb then we can use the prepare_for_registers simplification step to put the entrypoint back to the one that has the DMB (and remove the CFence)

That would require duplicating the entrypoints for arrays/non-arrays to add the with-DMB and not-with-DMB variant.

All quite straightforward but the line count does go up.

How should I proceed?

Patch-set: 1
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>