Update patch set 1
Patch Set 1:

(1 comment)

I'm not sure about so many 'auto', it makes it quite hard to follow

   auto x = foo->bar()->baz()
     .... = x->work()

The type information is just completely lost here. Is the guideline only use IDEs to read this code?

   container<TypeOfX>::iterator = foo->bar()->baz()
       ... = x->work

is much cleaner because it gives us the type of (foo->bar()), the typeof baz(), and the type of *x 

    foo->bar() is a container<TypeOfX>
    baz() returns the iterator
    *x is a TypeOfX

----

Patch-set: 1
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>