Testrunner: Fix concurrecy for --target
Popen was returning bytes instead of string which was causing the
logic of extracting out concurrency for the target device to break.
Now, the output of the Popen will be converted to string and regex
matching is used instead of string split to better handle the case
when the input string doesn't follow the pattern.

Test: ./art/test/testrunner/testrunner.py --target
Bug: 38130322

Change-Id: I6c2f3070e88e0cbd80a61380f6c4f714e8886b3c