ART: Add arena tracking mode
Add an arena tracking mode to get better insight into allocation
behavior. In this mode, the default size of arenas is very small
so that arena use for multiple places is limited. At the same time,
do not release arenas back into their pool and instead free them.

To limit the build-time impact, add an -inl file for the arena
allocator that defines the controlling flag and the default arena
size.

Bug: 34053922
Test: m test-art-host
Change-Id: I09bb5e743d7dc47e499a402d6fcac637c16a26ad