Do not cache profiles in the ProfileSaver.
Profiles with a lot of inline caches have a large memory
footprint. Do not cache them in the saver anymore.

The cache was useful to avoid doing unnecessary IO but it
takes too much memory now. Disable it until we update the
profile format to take less space.

This trades off IO for memory. As an effect we will gonna
do one extra read to detect if we have enough new information
worth saving. The saving period have been increased to 40secs
(from 20) in a previous CL which also helps to balance the IO.

Test: m test-art-host-gtest
      manual inspection with meminfo
Bug: 37711886

Change-Id: I9cfdd6fb70c289221e74ccf1b6449f28a7fb693d