Update patch set 5
Patch Set 5:

I think the behavior is correct? The tool doesn't rely on that field for the backtrace size and ignores the num member (which happens to always be one currently). Anyways, my changes are mostly for the html UI, are you OK with submitting?

Patch-set: 5
Reviewer: Gerrit User 1015378 <1015378@85c56323-6fa9-3386-8a01-6480fb634889>