Update patch set 6
Patch Set 6:

> I think the behavior is correct? The tool doesn't rely on that
 > field for the backtrace size and ignores the num member (which
 > happens to always be one currently). Anyways, my changes are mostly
 > for the html UI, are you OK with submitting?

This may work for now, but if the debug malloc implementation changes to make use of the 'num' field, this script will silently return wrong results.

Why have such a tool be in art instead of in bionic where debug malloc is implemented or as part of Android Studio? Whose job is it to test and maintain the tool? Who are the intended users of the tool? (this matters if we ever want to make changes to the tool interface.) I think these are important questions. I have not had time or priority to follow up on them.

Patch-set: 6
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>