Use GC critical section for VisitClasses
Since visit classes uses DecodeWeakGlobal, it may block on decoding
weak globals deadlock if the GC is running. This can cause deadlocks
since it will be blocked while holding the classlinker classes lock.

Using a GC critical section ensures that the GC is not running while
we visit the classes.

Bug: 38350590
Test: test-art-host

Change-Id: I1a529c4ec6b53c1979390741afc767d09a1faae0