Update patch set 4
Patch Set 4:

(4 comments)

> I would argue that having more info in hprof dumps is better than
 > less info. Is it bad if some of the linear alloc memory is
 > accounted for in the dump?

My concern with putting linear alloc memory in the dump is that developers implicitly assume everything shown as part of the 'app' heap (for example) in a heap dump takes up memory on the Java Heap. It could be surprising for them to learn that some of the byte arrays reported as part of the Java heap are really native allocations. I get a lot of questions from users trying to reconcile byte counts reported in heap dumps with page counts from meminfo.

I would suggest putting the overhead arrays in a new 'native' heap, but then you run into the problem that the app/zygote/image labels are currently used both to know what heap an object is allocated on and, just as important if not more so, to identify which bytes the developer can do something about. A single "native" heap doesn't convey the latter information. You would need "app_native", "zygote_native", and "image_native" for that, which starts to become tedious.

If we do include overhead info in the heap dump, it's also important to make that information actionable. For example, with this CL I see that conscrypt.NativeCrypto has almost 13K of overhead (compared to 224 bytes of class object), but I have no insight into what that overhead is from or how to reduce it. Can you separate the different kind of overheads, so there is a "method_overhead", "field_overhead", and "imt_overhead"? If I see a large field_overhead, I know to try reducing the number of fields in my class. If I see a large method_overhead, I know to try reducing the number of methods in my class.

Taking all of this into consideration, my vote is to continue to include the overhead from the linear alloc space in the heap dumps using the same heap label as the class object, but split the $overhead field into three different fields to distinguish between field, method, and imt overhead.

Patch-set: 4
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>