ART: Introduce fast-debug level
Add support for DebugCheckLevel::kFast vs DebugCheckLevel::kSlow to mark
debug checks that are slow. This allows to have a debug build without
expensive checks, which makes a debug build practical to be used on a device.

The default for a debug build will now be fast. Ensure that run-tests include
all debug checks, and add a check to SignalTest.

Current slow checks:
* Class status checking in ArtMethod
* Read barrier invariant checking

Update logging_test.

Bug: 35644369
Test: m test-art-host
Change-Id: I66138a9accc601fa0fa675cf600412ad3629d28f