Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> We should use Deallocate, in case we ever want to do something like CheckJNI. What's stopping you from using JvmtiUniquePtr?

We don't have any good idea of which one it is coming from since that is all hidden by the event_handler.

I was thinking we could leave the bug open and have this as a bit of a temporary fix until we can come up with something better.

Patch-set: 1
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>