Update patch set 1
Patch Set 1:

Not sure about this, it's a bit like writing code that has strong exception guarantee (i.e. the point is unclear).

Yes, it can be done, but we compile with -fno-exceptions so it would've just crashed anyway if there was an exception.

I posted in the bug about this.

Patch-set: 1
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>