An InstrumentationListener to profile interpreter
Added a new CLInstrumentationListener to count how many instructions are
executed by interpreter and a runtime option '-Xprofile-interpreter' to
enable this profiler. It's disabled by default becuase it has huge
impact on interpreter's performance.

The listener is registered when Zygote is forked and dump the result
when SIGQUIT is received for dumping heap information.

Another atomic_int is used for counting how many native functions are
called from interpreter.

Test: test-art-host -j64
Change-Id: Ib105fbbced7c87d53d80d325af21c6b2d619076a