An InstrumentationListener to profile interpreter
Added a new CLInstrumentationListener to count how many instructions are
executed by interpreter and a runtime option '-Xprofile-interpreter' to
enable this profiler. It's disabled by default becuase it has huge
impact on interpreter's performance.

The listener is registered when Zygote is forked and dump the result
when SIGQUIT is received for dumping heap information.

Test: art-test-host -j64
Change-Id: Ib105fbbced7c87d53d80d325af21c6b2d619076a