ART: Move source owner data to LogContentionEvent
Change the signature of Monitor::LogContentionEvent to take the
owner's ArtMethod and dex pc instead of the derived data. Move the
decoding from monitor.cc to monitor_android.cc.

This change does not change functionality. It is only in preparation
for a later, actual change of what is being logged.

Bug: 62241642
Test: m test-art-host
Change-Id: I87b10f6f31f5ba3c177d04dd63ad54167ab9dde5